Introducing new PhotoSystem object in the photometry submodule to interface with Isochrone objects
Isochrone objects used to serve 2 roles. The first was the intended purpose, to represent the on-disk data of a given photometric system (including descriptors and single isochrone files). But the submodule implementation eventually forced those objects to assume a 2nd role to host various methods that helped the package as a whole in how it would exploit the photometric system characteritics. This 2nd role should have been given to another class in the first place, which PhotoSystem will do from now on.